+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/component/textfield/PasswordFieldIT.java
@@ -3,7 +3,7 @@ package junit.com.vaadin.testbench.tests.component.textfield;
 import com.vaadin.flow.component.textfield.testbench.PasswordFieldElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import junit.com.vaadin.testbench.tests.component.common.AbstractIT;
-import junit.com.vaadin.testbench.tests.testui.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 
 import static com.vaadin.flow.component.textfield.testbench.test.PasswordFieldView.INITIAL_VALUE;
