+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/component/upload/UploadIT.java
@@ -4,7 +4,7 @@ import com.vaadin.flow.component.upload.testbench.UploadElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import com.vaadin.testbench.addons.webdriver.SkipBrowsers;
 import junit.com.vaadin.testbench.tests.component.common.AbstractIT;
-import junit.com.vaadin.testbench.tests.testui.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
