+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/uitest/PerformanceTest.java
@@ -1,7 +1,7 @@
-package junit.com.vaadin.testbench.tests.testui;
+package junit.com.vaadin.testbench.tests.uitest;
 
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
-import junit.com.vaadin.testbench.tests.testui.elements.NativeButtonElement;
+import junit.com.vaadin.testbench.tests.uitest.elements.NativeButtonElement;
 import org.junit.jupiter.api.Assertions;
 
 import static com.vaadin.testbench.tests.testUI.PerformanceView.ROUTE;
