+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/uitest/SVGTest.java
@@ -1,4 +1,4 @@
-package junit.com.vaadin.testbench.tests.testui;
+package junit.com.vaadin.testbench.tests.uitest;
 
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import com.vaadin.testbench.addons.webdriver.SkipBrowsers;
