+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/uitest/elements/ExecuteJavascriptTest.java
@@ -1,8 +1,8 @@
-package junit.com.vaadin.testbench.tests.testui.elements;
+package junit.com.vaadin.testbench.tests.uitest.elements;
 
 import com.vaadin.testbench.TestBenchElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
-import junit.com.vaadin.testbench.tests.testui.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 
 import static com.vaadin.testbench.tests.testUI.ElementQueryView.ROUTE;
